backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
ffmpeg-python>=0.2.0
openai>=1.6.0
oss2>=2.18.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
aiofiles>=23.2.1
